break out undo/repo workflow editor tests into their own package
